main.py
import asyncio
import logging
import os
from aiogram import Bot, Dispatcher
from dotenv import load_dotenv
from survey import router as survey_router
from menu import router as menu_router
from callbacks import router as callback_router

load_dotenv()
TOKEN = os.getenv("TOKEN")

bot = Bot(token=TOKEN)
dp = Dispatcher()

logging.basicConfig(level=logging.INFO)

dp.include_router(survey_router)
dp.include_router(menu_router)
dp.include_router(callback_router)

async def main():
    print("Starting bot...")
    await dp.start_polling(bot)

if __name__ == "__main__":
    asyncio.run(main())