requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
asyncpg==0.30.0
python-telegram-bot==22.0
aiohttp==3.11.14
beautifulsoup4~=4.13.3